created new feature for foote and developed some methods and page objects
