--- src/test/java/com/test/mirchi/modules/HeaderNavAction.java
@@ -1,24 +1,0 @@
-package com.test.mirchi.modules;
-
-import java.util.List;
-
-import org.junit.Assert;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Reporter;
-
-import com.test.mirchi.pageobjects.MirchiHomePage;
-
-public class HeaderNavAction {
-	public static void Execute(WebDriver driver) throws Exception {
-		WebDriverWait wait = new WebDriverWait(driver, 10);
-		List<WebElement> items = driver.findElements(By.className("menu-item"));
-		Reporter.log("Found Menu links: ", items.size());
-		Assert.assertEquals(5,items.size());
-		MirchiHomePage.HeaderPage.menu_item.click();;
-
-		Reporter.log("Select product successful");
-	}
-}
