--- src/test/java/com/test/mirchi/step_definitions/HeaderNavSteps.java
@@ -1,28 +1,0 @@
-package com.test.mirchi.step_definitions;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.PageFactory;
-
-import com.test.mirchi.modules.HeaderNavAction;
-import com.test.mirchi.pageobjects.MirchiHomePage;
-
-import cucumber.api.java.en.Then;
-
-public class HeaderNavSteps {
-
-    public WebDriver driver;
-    
-
-    public HeaderNavSteps()
-    {
-        driver = Hooks.driver;
-        
-    }
-    
-    @Then("^I select a links and verify if works and lands on right page$")
-    public void i_select_a_links_and_verify_if_works_and_lands_on_right_page() throws Throwable {
-    	PageFactory.initElements(driver,MirchiHomePage.FooterPage.class);
-    	HeaderNavAction.Execute(driver);
-    }
-
-}
